SC_BasicConfig.yamlのtakao_baseの設定に対応
