+++ firmware/src/DebugTools.cpp
@@ -1,0 +1,40 @@
+#include <M5Unified.h>
+#include "DebugTools.h"
+
+
+void check_heap_free_size(void){
+  Serial.printf("===============================================================\n");
+  Serial.printf("Check free heap size\n");
+  Serial.printf("===============================================================\n");
+  //Serial.printf("esp_get_free_heap_size()                              : %6d\n", esp_get_free_heap_size() );
+  Serial.printf("heap_caps_get_free_size(MALLOC_CAP_DMA)               : %6d\n", heap_caps_get_free_size(MALLOC_CAP_DMA) );
+  Serial.printf("heap_caps_get_free_size(MALLOC_CAP_SPIRAM)            : %6d\n", heap_caps_get_free_size(MALLOC_CAP_SPIRAM) );
+  Serial.printf("heap_caps_get_free_size(MALLOC_CAP_INTERNAL)          : %6d\n", heap_caps_get_free_size(MALLOC_CAP_INTERNAL) );
+  Serial.printf("heap_caps_get_free_size(MALLOC_CAP_DEFAULT)           : %6d\n", heap_caps_get_free_size(MALLOC_CAP_DEFAULT) );
+
+}
+
+void check_heap_largest_free_block(void){
+  Serial.printf("===============================================================\n");
+  Serial.printf("Check largest free heap block\n");
+  Serial.printf("===============================================================\n");
+  Serial.printf("heap_caps_get_largest_free_block(MALLOC_CAP_DMA)      : %6d\n", heap_caps_get_largest_free_block(MALLOC_CAP_DMA) );
+  Serial.printf("heap_caps_get_largest_free_block(MALLOC_CAP_SPIRAM)   : %6d\n", heap_caps_get_largest_free_block(MALLOC_CAP_SPIRAM) );
+  Serial.printf("heap_caps_get_largest_free_block(MALLOC_CAP_INTERNAL) : %6d\n", heap_caps_get_largest_free_block(MALLOC_CAP_INTERNAL) );
+  Serial.printf("heap_caps_get_largest_free_block(MALLOC_CAP_DEFAULT)  : %6d\n", heap_caps_get_largest_free_block(MALLOC_CAP_DEFAULT) );
+
+}
+
+uint32_t get_elapsed_time_micro(const char* str){
+  static uint32_t prev = 0;
+  uint32_t now = 0;
+  uint32_t elapsedTime = 0;
+  now = micros();
+  if(prev == 0){
+    prev = now;
+  }
+  elapsedTime = now - prev;
+  prev = now;
+  Serial.printf("%s: %d [us]\n", str, elapsedTime);
+  return elapsedTime;
+}
